Split __init__ into _filters and _utils
